refactor(within): narrow crate root re-exports for stable API
Move lower-level types out of the crate root so the advertised stable
surface is just the entry points, params, errors, and `FactorMajorStore`.
Items remain accessible via their module paths
(`within::operator::gramian::Gramian`, `within::config::LocalSolverConfig`,
etc.).

Tests, benches, examples, and the `solver.rs` doctest are updated to use
module paths. Python API is unchanged; `within-py` already imported via
`within::config::*`, `within::domain::Design`, and
`within::observation::FactorMajorStore`.
