+++ crates/within/tests/domain.rs
@@ -5,9 +5,10 @@
 
 use proptest::prelude::*;
 use schwarz_precond::Operator;
+use within::domain::Design;
 use within::observation::FactorMajorStore;
 use within::operator::gramian::Gramian;
-use within::{Design, DesignOperator};
+use within::operator::DesignOperator;
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -270,7 +271,8 @@ proptest! {
 
 #[test]
 fn test_three_factor_design_solve_converges() {
-    use within::{solve, LocalSolverConfig, Preconditioner, SolverParams};
+    use within::config::LocalSolverConfig;
+    use within::{solve, Preconditioner, SolverParams};
 
     let n_obs = 60;
     let n_lev = 5usize;
@@ -326,7 +328,8 @@ fn test_three_factor_design_solve_converges() {
 /// to cover more of the preconditioner path.
 #[test]
 fn test_three_factor_design_multiplicative_schwarz_converges() {
-    use within::{solve, LocalSolverConfig, Preconditioner, SolverParams};
+    use within::config::LocalSolverConfig;
+    use within::{solve, Preconditioner, SolverParams};
 
     let n_obs = 60;
     let n_lev = 5usize;
@@ -384,7 +387,8 @@ fn test_three_factor_design_multiplicative_schwarz_converges() {
 /// 2 subdomains — correctness is validated indirectly through convergence.
 #[test]
 fn test_disconnected_design_larger_converges() {
-    use within::{solve, LocalSolverConfig, Preconditioner, SolverParams};
+    use within::config::LocalSolverConfig;
+    use within::{solve, Preconditioner, SolverParams};
 
     // Extend the disconnected example to more observations so the solve is
     // non-trivial: component A has factor-0 levels {0,1}, factor-1 levels {0,1,2};
@@ -430,7 +434,8 @@ fn test_disconnected_design_larger_converges() {
 
 #[test]
 fn test_disconnected_design_solve_converges() {
-    use within::{solve, LocalSolverConfig, Preconditioner, SolverParams};
+    use within::config::LocalSolverConfig;
+    use within::{solve, Preconditioner, SolverParams};
 
     let n_obs = 4;
     let mut cats = ndarray::Array2::<u32>::zeros((n_obs, 2));
